Thêm nhiều chức năng vcl, sr vì t đợi xong hết mới push, Commit chi tiết vào nào  - Sửa lại model, DB, thêm CurrencyId cho transcation và Cheque, thêm vào newBalancesender và NewBalanceRecipient  - Sửa lại các cái liên quan đến tiền thành double  - Thiết kế lại toàn bộ trang SendIndex gồm:   + Thêm danh sách người nhận (tạm thời t chưa cho gửi cho người nhận mới)   + thêm Danh sách tài khoản muốn dùng để chuyển khoản   + Dùng các Currency nhìn cho chuyên cmn nghiệp   + Thêm javascript để tính toán đổi tiền tự động (JS t nhét thẳng vào cshtml cho dễ nhớ)   + Xóa cái fee đi, khó chịu vcl
- Sửa lại SendController gồm:
  + Thêm vào các Viewbag cho các action của Send: SendIndex, SendCofirm, SendSuccess
  + Hoàn tất quy trình chuyển tiền, đã cập nhật được cho Db.

+++ OnlineBanking/Models/Account.cs
@@ -24,7 +24,7 @@ namespace OnlineBanking.Models
         [DisplayName("Account Number")]
         [Required(ErrorMessage = ("{0} Account Number Cannot Be Blank"))]
         public string Number { get; set; }
-        public int Balance { get; set; }
+        public double Balance { get; set; }
         public DateTime CreateDate { get; set; }
 
         public bool Status { get; set; }
